Consolidate remember device specs
**Why**: So that as we add remember device for more MFA methods we can grow the testing suite in a reasonable way
