+++ spec/features/remember_device/sp_expiration_spec.rb
